+++ internal/diag.go
@@ -1,4 +1,4 @@
-package diag
+package internal
 
 import (
 	"go/ast"
@@ -14,7 +14,8 @@ func NewEmbeddedTypeAfterNotEmbeddedTypeDiag(embeddedField *ast.Field) analysis.
 }
 
 func NewMissingSpaceBetweenLastEmbeddedTypeAndFirstNotEmbeddedTypeDiag(
-	lastEmbeddedField *ast.Field) analysis.Diagnostic {
+	lastEmbeddedField *ast.Field,
+) analysis.Diagnostic {
 	return analysis.Diagnostic{
 		Pos:     lastEmbeddedField.Pos(),
 		Message: "there must be an empty line separating embedded fields from regular fields",